fix(m0-http): migrate the server/event-loop paths to nightly 2026-07-28
The previous commit made all 470 package tests pass, but the unit tests
never instantiate the socket, event-loop, or URI-parsing code, so those
paths were still broken. `mojo build apps/demo_tasks/server.mojo` now
succeeds, and the built server serves GET / and the GRAIL /conditions
endpoint correctly.

- `Addr` is now a linear trait — a trait that neither conforms to
  `ImplicitlyDeletable` nor is `@explicit_destroy` can no longer be
  implicitly destroyed, which broke Socket teardown/accept/bind. Added
  the conformance.
- `UnsafePointer(unsafe_from_address=0)` is now rejected outright for a
  literal null. The kevent() FFI signature models its nullable
  changelist/eventlist/timeout params as `OptionalUnsafePointer` and
  passes `None`; Optional[UnsafePointer] has the same layout with None as
  the null niche, so the ABI is unchanged. Remaining null *checks* compare
  `Int(ptr)` rather than using the removed `Bool(ptr)`.
- Interior origins: `String.as_bytes()` slices now carry the string's
  interior origin, so URI.parse's ByteView declaration was updated.
- `len(String/StringSlice)` -> `byte_length()` across address, request,
  uri, server, and event_loop.
- Remaining gated pointer ops switched to the ungated `unsafe_*` spellings.
- `List.capacity` is a method; `sleep` takes Int/Float64; setsockopt's
  `level` needs an explicit Int32; slot_header_start/DRAIN_TIMEOUT_NS
  are Int now that `perf_counter_ns()` returns Int; request_cookie_jar
  imports `Header` explicitly.

+++ packages/m0-http/lightbug_http/connection.mojo
@@ -287,7 +287,7 @@ struct ListenConfig:
                     socket.shutdown()
                 except shutdown_err:
                     pass
-                sleep(UInt(1))
+                sleep(1)
 
         try:
             socket.listen(128)
